fixture move to conftest and use explicitly
