+++ src/gloci/oras/registry.py
@@ -536,7 +536,6 @@ class GlociRegistry(Registry):
             info_data = yaml.safe_load(f)
             base_path = os.path.join(os.path.dirname(info_yaml))
         
-        features = read_feature_files("./features") 
         manifest_image = oras.oci.NewManifest()
         total_size = 0
