+++ tests/test_push_image.py
@@ -2,54 +2,13 @@ import subprocess
 import io
 from dotenv import load_dotenv
 import pytest
-import tempfile
 from click.testing import CliRunner
 from gloci.cli import cli
-from .helper import spawn_background_process
-import os
-import shutil
-import json
-import queue
-import threading
-import sys
 
 CONTAINER_NAME_ZOT_EXAMPLE = "127.0.0.1:18081/examplecontainer2"
 
 
-def write_zot_config(config_dict, file_path):
-    with open(file_path, "w") as config_file:
-        json.dump(config_dict, config_file, indent=4)
-
-
-def setup_test_environment():
-    zot_config = {
-        "distSpecVersion": "1.1.0",
-        "storage": {"rootDirectory": "output/registry/zot"},
-        "http": {"address": "127.0.0.1", "port": "18081"},
-        "log": {"level": "warn"},
-    }
-
-    with tempfile.NamedTemporaryFile(delete=False, suffix=".json") as temp_config_file:
-        write_zot_config(zot_config, temp_config_file.name)
-        zot_config_file_path = temp_config_file.name
-
-    print(f"Spawning zot registry with config {zot_config_file_path}")
-    zot_process = spawn_background_process(
-        f"zot serve {zot_config_file_path}",
-        stdout=sys.stdout,
-        stderr=sys.stderr,
-    )
-
-    yield zot_process
-
-    zot_process.terminate()
-
-    if os.path.isdir("./output"):
-        shutil.rmtree("./output")
-    if os.path.isfile(zot_config_file_path):
-        os.remove(zot_config_file_path)
-
-
+@pytest.mark.usefixtures('zot_session')
 @pytest.mark.parametrize(
     "info_yaml_path, version, cname, arch",
     [
